[#13] Feat InformationPage - 파일구조 변경
