+++ src/components/index.tsx
@@ -4,3 +4,4 @@ export * from "./modal/alert";
 export * from "./modal/modal";
 export * from "./loginPage/LoginPage";
 export * from "./signUpPage/SignUpPage";
+export * from "./infoPage/InformationPage";
