+++ src/components/infoPage/Geumju.tsx
@@ -1,24 +1,22 @@
 import React from "react";
 import styled from "styled-components";
-import { customColor } from "../components/customColor";
+import { customColor } from "../customColor";
 import Image from "next/image";
-import email from "../assets/img/information/email.png";
-import phone from "../assets/img/information/phone.png";
-import imac from "../assets/img/information/imac.png";
-import character from "../assets/img/information/character.png"
-const Cheongjo = () => {
-
-  
+import email from "../../assets/img/information/email.png"
+import phone from "../../assets/img/information/phone.png";
+import imac from "../../assets/img/information/imac.png";
+import character from "../../assets/img/information/character.png"
+const Geumju = () => {
   return (
     <Wrapper>
       <InfoBox>
-      <Circle>
+        <Circle>
           <Image src={character} alt = {''} width={80} />
         </Circle>
-        <Name>박 청 조</Name>
+        <Name>조 금 주</Name>
         <TaskBox>
           <Image src={imac} alt={""} width={35} />
-          <Task>Back</Task>
+          <Task>Design</Task>
         </TaskBox>
         <NumberBox>
           <Image src={phone} alt={""} width={20} />
@@ -26,19 +24,18 @@ const Cheongjo = () => {
         </NumberBox>
         <EmailBox>
           <Image src={email} alt={""} width={20} />
-          <Email>aaaaaaa@naver.com</Email>
+          <Email>aaaaaaaaa@naver.com</Email>
         </EmailBox>
       </InfoBox>
     </Wrapper>
   );
 };
 
-export default Cheongjo;
+export default Geumju;
 
 const Wrapper = styled.div`
   position: relative;
   display: flex;
-  
 `;
 const InfoBox = styled.div`
   width: 150px;
