+++ src/components/infoPage/Hyoseong.tsx
@@ -1,12 +1,14 @@
 import React from "react";
 import styled from "styled-components";
-import { customColor } from "../components/customColor";
+import { customColor } from "../customColor";
 import Image from "next/image";
-import email from "../assets/img/information/email.png";
-import phone from "../assets/img/information/phone.png";
-import imac from "../assets/img/information/imac.png";
-import character from "../assets/img/information/character.png"
+import email from "../../assets/img/information/email.png"
+import phone from "../../assets/img/information/phone.png";
+import imac from "../../assets/img/information/imac.png";
+import character from "../../assets/img/information/character.png"
 const Hyoseong = () => {
+
+
   return (
     <Wrapper>
       <InfoBox>
@@ -36,6 +38,9 @@ export default Hyoseong;
 const Wrapper = styled.div`
   position: relative;
   display: flex;
+  &:hover{  
+    cursor: pointer;
+  }
 `;
 const InfoBox = styled.div`
   width: 150px;
